Fix https://github.com/ultralytics/yolov5/issues/8405

+++ detect/utils/general.py
@@ -553,7 +553,7 @@ def build_targets(p, targets, model):
 
         # Append
         a = t[:, 6].long()  # anchor indices
-        indices.append((b, a, gj.clamp_(0, gain[3]), gi.clamp_(0, gain[2])))  # image, anchor, grid indices
+        indices.append((b, a, gj.clamp_(0, gain[3].int()), gi.clamp_(0, gain[2].int())))  # image, anchor, grid indices
         tbox.append(torch.cat((gxy - gij, gwh), 1))  # box
         anch.append(anchors[a])  # anchors
         tcls.append(c)  # class
